+++ src/StackAdmin/Resources/Commands.Resources.Test/ScenarioTests/ResourcesController.cs
@@ -1,0 +1,339 @@
+﻿// ----------------------------------------------------------------------------------
+//
+// Copyright Microsoft Corporation
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ----------------------------------------------------------------------------------
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Net.Http;
+using System.Net.Http.Headers;
+using Microsoft.Azure.Commands.Common.Authentication;
+using Microsoft.Azure.Commands.ResourceManager.Cmdlets.Components;
+using Microsoft.Azure.Commands.ResourceManager.Cmdlets.Extensions;
+using Microsoft.Azure.Graph.RBAC.Version1_6;
+using Microsoft.Azure.Insights;
+using Microsoft.Azure.Management.Authorization.Version2015_07_01;
+using Microsoft.Azure.Management.ResourceManager;
+using Microsoft.Azure.Test.HttpRecorder;
+using Microsoft.Rest.ClientRuntime.Azure.TestFramework;
+using Microsoft.WindowsAzure.Commands.Common;
+using Microsoft.WindowsAzure.Commands.ScenarioTest;
+using LegacyRMClient = Microsoft.Azure.Management.Resources;
+using LegacyRMSubscription = Microsoft.Azure.Subscriptions;
+using LegacyTest = Microsoft.Azure.Test;
+using TestEnvironmentFactory = Microsoft.Rest.ClientRuntime.Azure.TestFramework.TestEnvironmentFactory;
+using TestUtilities = Microsoft.Rest.ClientRuntime.Azure.TestFramework.TestUtilities;
+using Microsoft.Azure.Test.Authentication;
+using Microsoft.Azure.ServiceManagemenet.Common.Models;
+using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
+using Microsoft.Azure.Commands.Resources.Models.Gallery;
+
+namespace Microsoft.Azure.Commands.Resources.Test.ScenarioTests
+{
+    public sealed class ResourcesController
+    {
+        private LegacyTest.CSMTestEnvironmentFactory csmTestFactory;
+        private EnvironmentSetupHelper helper;
+        private const string TenantIdKey = "TenantId";
+        private const string DomainKey = "Domain";
+        private const string SubscriptionIdKey = "SubscriptionId";
+
+        public GraphRbacManagementClient GraphClient { get; private set; }
+
+        public ResourceManagementClient ResourceManagementClient { get; private set; }
+
+        public LegacyRMClient.ResourceManagementClient LegacyResourceManagementClient { get; private set; }
+
+        public LegacyRMSubscription.SubscriptionClient LegacySubscriptionClient { get; private set; }
+
+        public FeatureClient FeatureClient { get; private set; }
+
+        public Internal.Subscriptions.SubscriptionClient SubscriptionClient { get; private set; }
+
+        public GalleryClient GalleryClient { get; private set; }
+
+        public InsightsClient InsightsClient { get; private set; }
+
+        public AuthorizationManagementClient AuthorizationManagementClient { get; private set; }
+
+
+        public string UserDomain { get; private set; }
+
+        public static ResourcesController NewInstance
+        {
+            get
+            {
+                return new ResourcesController();
+            }
+        }
+
+        public ResourcesController()
+        {
+            helper = new EnvironmentSetupHelper();
+        }
+
+        public void RunPsTest(params string[] scripts)
+        {
+            var callingClassType = TestUtilities.GetCallingClass(2);
+            var mockName = TestUtilities.GetCurrentMethodName(2);
+
+            RunPsTestWorkflow(
+                () => scripts,
+                // no custom initializer
+                null,
+                // no custom cleanup 
+                null,
+                callingClassType,
+                mockName);
+        }
+
+        public void RunPsTestWorkflow(
+            XunitTracingInterceptor interceptor,
+            Func<string[]> scriptBuilder,
+            Action<LegacyTest.CSMTestEnvironmentFactory> initialize,
+            Action cleanup,
+            string callingClassType,
+            string mockName)
+        {
+            helper.TracingInterceptor = interceptor;
+            this.RunPsTestWorkflow(scriptBuilder, initialize, cleanup, callingClassType, mockName);
+        }
+
+        public void RunPsTestWorkflow(
+            Func<string[]> scriptBuilder,
+            Action<LegacyTest.CSMTestEnvironmentFactory> initialize,
+            Action cleanup,
+            string callingClassType,
+            string mockName)
+        {
+            Dictionary<string, string> d = new Dictionary<string, string>();
+            d.Add("Microsoft.Resources", null);
+            d.Add("Microsoft.Features", null);
+            d.Add("Microsoft.Authorization", null);
+            var providersToIgnore = new Dictionary<string, string>();
+            providersToIgnore.Add("Microsoft.Azure.Management.ResourceManager.ResourceManagementClient", "2016-07-01");
+            providersToIgnore.Add("Microsoft.Azure.Management.Resources.ResourceManagementClient", "2016-02-01");
+            HttpMockServer.Matcher = new ResourcesRecordMatcher(true, d, providersToIgnore);
+            HttpMockServer.RecordsDirectory = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "SessionRecords");
+
+            using (MockContext context = MockContext.Start(callingClassType, mockName))
+            {
+                this.csmTestFactory = new LegacyTest.CSMTestEnvironmentFactory();
+
+                if (initialize != null)
+                {
+                    initialize(this.csmTestFactory);
+                }
+
+                helper.SetupEnvironment(AzureModule.AzureResourceManager);
+
+                SetupManagementClients(context);
+
+                var callingClassName = callingClassType
+                                        .Split(new[] { "." }, StringSplitOptions.RemoveEmptyEntries)
+                                        .Last();
+                helper.SetupModules(AzureModule.AzureResourceManager,
+                    "ScenarioTests\\Common.ps1",
+                    "ScenarioTests\\" + callingClassName + ".ps1",
+                    helper.StackRMProfileModule,
+                    helper.StackRMResourceModule,
+                    helper.StackRMStorageModule);
+
+                try
+                {
+                    if (scriptBuilder != null)
+                    {
+                        var psScripts = scriptBuilder();
+
+                        if (psScripts != null)
+                        {
+                            helper.RunPowerShellTest(psScripts);
+                        }
+                    }
+                }
+                finally
+                {
+                    if (cleanup != null)
+                    {
+                        cleanup();
+                    }
+                }
+            }
+        }
+
+        private void SetupManagementClients(MockContext context)
+        {
+            LegacyResourceManagementClient = GetLegacyResourceManagementClient();
+            LegacySubscriptionClient = GetLegacySubscriptionClient();
+            ResourceManagementClient = GetResourceManagementClient(context);
+            SubscriptionClient = GetSubscriptionClient(context);
+            GalleryClient = GetGalleryClient();
+            AuthorizationManagementClient = GetAuthorizationManagementClient(context);
+            GraphClient = GetGraphClient(context);
+            InsightsClient = GetInsightsClient();
+            this.FeatureClient = this.GetFeatureClient(context);
+            var testEnvironment = this.csmTestFactory.GetTestEnvironment();
+            var credentials = new SubscriptionCredentialsAdapter(
+                testEnvironment.AuthorizationContext.TokenCredentials[Microsoft.Azure.Test.TokenAudience.Management],
+                testEnvironment.SubscriptionId);
+            HttpClientHelperFactory.Instance = new TestHttpClientHelperFactory(credentials);
+
+            helper.SetupManagementClients(ResourceManagementClient,
+                LegacyResourceManagementClient,
+                LegacySubscriptionClient,
+                SubscriptionClient,
+                GalleryClient,
+                AuthorizationManagementClient,
+                GraphClient,
+                InsightsClient,
+                this.FeatureClient);
+        }
+
+        private GraphRbacManagementClient GetGraphClient(MockContext context)
+        {
+            var environment = TestEnvironmentFactory.GetTestEnvironment();
+            string tenantId = null;
+
+            if (HttpMockServer.Mode == HttpRecorderMode.Record)
+            {
+                tenantId = environment.Tenant;
+                UserDomain = String.IsNullOrEmpty(environment.UserName) ? String.Empty : environment.UserName.Split(new[] { "@" }, StringSplitOptions.RemoveEmptyEntries).Last();
+
+                HttpMockServer.Variables[TenantIdKey] = tenantId;
+                HttpMockServer.Variables[DomainKey] = UserDomain;
+            }
+            else if (HttpMockServer.Mode == HttpRecorderMode.Playback)
+            {
+                if (HttpMockServer.Variables.ContainsKey(TenantIdKey))
+                {
+                    tenantId = HttpMockServer.Variables[TenantIdKey];
+                }
+                if (HttpMockServer.Variables.ContainsKey(DomainKey))
+                {
+                    UserDomain = HttpMockServer.Variables[DomainKey];
+                }
+                if (HttpMockServer.Variables.ContainsKey(SubscriptionIdKey))
+                {
+                    AzureRmProfileProvider.Instance.Profile.DefaultContext.Subscription.Id = HttpMockServer.Variables[SubscriptionIdKey];
+                }
+            }
+
+            var client = context.GetGraphServiceClient<GraphRbacManagementClient>(environment);
+            client.TenantID = tenantId;
+            if (AzureRmProfileProvider.Instance != null &&
+                AzureRmProfileProvider.Instance.Profile != null &&
+                AzureRmProfileProvider.Instance.Profile.DefaultContext != null &&
+                AzureRmProfileProvider.Instance.Profile.DefaultContext.Tenant != null)
+            {
+                AzureRmProfileProvider.Instance.Profile.DefaultContext.Tenant.Id = client.TenantID;
+            }
+            return client;
+        }
+
+        private AuthorizationManagementClient GetAuthorizationManagementClient(MockContext context)
+        {
+            return context.GetServiceClient<AuthorizationManagementClient>(TestEnvironmentFactory.GetTestEnvironment());
+        }
+
+        private FeatureClient GetFeatureClient(MockContext context)
+        {
+            return context.GetServiceClient<FeatureClient>(TestEnvironmentFactory.GetTestEnvironment());
+        }
+
+        private ResourceManagementClient GetResourceManagementClient(MockContext context)
+        {
+            return context.GetServiceClient<ResourceManagementClient>(TestEnvironmentFactory.GetTestEnvironment());
+        }
+
+        private LegacyRMClient.ResourceManagementClient GetLegacyResourceManagementClient()
+        {
+            return LegacyTest.TestBase.GetServiceClient<LegacyRMClient.ResourceManagementClient>(this.csmTestFactory);
+        }
+
+        private LegacyRMSubscription.SubscriptionClient GetLegacySubscriptionClient()
+        {
+            return LegacyTest.TestBase.GetServiceClient<LegacyRMSubscription.SubscriptionClient>(this.csmTestFactory);
+        }
+
+        private Internal.Subscriptions.SubscriptionClient GetSubscriptionClient(MockContext context)
+        {
+            return context.GetServiceClient<Internal.Subscriptions.SubscriptionClient>(TestEnvironmentFactory.GetTestEnvironment());
+        }
+
+        private GalleryClient GetGalleryClient()
+        {
+            return LegacyTest.TestBase.GetServiceClient<GalleryClient>(this.csmTestFactory);
+        }
+
+        private InsightsClient GetInsightsClient()
+        {
+            return LegacyTest.TestBase.GetServiceClient<InsightsClient>(this.csmTestFactory);
+        }
+
+        /// <summary>
+        /// The test http client helper factory.
+        /// </summary>
+        private class TestHttpClientHelperFactory : HttpClientHelperFactory
+        {
+            /// <summary>
+            /// The subscription cloud credentials.
+            /// </summary>
+            private readonly SubscriptionCloudCredentials credential;
+
+            /// <summary>
+            /// Initializes a new instance of the <see cref="TestHttpClientHelperFactory"/> class.
+            /// </summary>
+            /// <param name="credentials"></param>
+            public TestHttpClientHelperFactory(SubscriptionCloudCredentials credentials)
+            {
+                this.credential = credentials;
+            }
+
+            /// <summary>
+            /// Creates new instances of the <see cref="HttpClientHelper"/> class.
+            /// </summary>
+            /// <param name="credentials">The credentials.</param>
+            /// <param name="headerValues">The headers.</param>
+            public override HttpClientHelper CreateHttpClientHelper(SubscriptionCloudCredentials credentials, IEnumerable<ProductInfoHeaderValue> headerValues, Dictionary<string, string> cmdletHeaderValues)
+            {
+                return new HttpClientHelperImpl(credentials: this.credential, headerValues: headerValues, cmdletHeaderValues: cmdletHeaderValues);
+            }
+
+            /// <summary>
+            /// An implementation of the <see cref="HttpClientHelper"/> abstract class.
+            /// </summary>
+            private class HttpClientHelperImpl : HttpClientHelper
+            {
+                /// <summary>
+                /// Initializes new instances of the <see cref="HttpClientHelperImpl"/> class.
+                /// </summary>
+                /// <param name="credentials">The credentials.</param>
+                /// <param name="headerValues">The headers.</param>
+                public HttpClientHelperImpl(SubscriptionCloudCredentials credentials, IEnumerable<ProductInfoHeaderValue> headerValues, Dictionary<string, string> cmdletHeaderValues)
+                    : base(credentials: credentials, headerValues: headerValues, cmdletHeaderValues: cmdletHeaderValues)
+                {
+                }
+
+                /// <summary>
+                /// Creates an <see cref="HttpClient"/>
+                /// </summary>
+                /// <param name="primaryHandlers">The handlers that will be added to the top of the chain.</param>
+                public override HttpClient CreateHttpClient(params DelegatingHandler[] primaryHandlers)
+                {
+                    return base.CreateHttpClient(HttpMockServer.CreateInstance().AsArray().Concat(primaryHandlers).ToArray());
+                }
+            }
+        }
+    }
+}