+++ src/ResourceManager/DataMigration/Commands.DataMigration/ChangeLog.md
@@ -18,6 +18,12 @@
         - Additional information about change #1
 -->
 ## Current Release
+
+## Version 0.1.6
+* Fixed issue with default resource groups not being set.
+* Updated common runtime assemblies
+
+## Version 0.1.5
 * Fixed issue with default resource groups not being set.
 
 ## Version 0.1.4