+++ src/ResourceManager/Storage/Commands.Management.Storage/ChangeLog.md
@@ -18,6 +18,13 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Upgrade to Azure Storage Client Library 9.3.0 
+
+## Version 5.0.4
+* Fixed issue with default resource groups not being set.
+* Updated common runtime assemblies
+
+## Version 5.0.3
 * Fixed issue with default resource groups not being set.
 
 ## Version 5.0.2